refactor: eslint
