+++ async-race/src/components/api/api.ts
@@ -1,11 +1,7 @@
-import { ICar } from './../shared/interfaces';
-type JSONResponse = {
-  items: object[],
-  count: string | null
-}
+import { ICar } from '../shared/interfaces';
 
 export class Api {
-  cars: object[] = [];
+  cars = [];
 
   private readonly base = 'http://localhost:3000';
 
@@ -15,17 +11,13 @@ export class Api {
 
   private readonly winnres = `${this.base}/winnres`;
 
-  async getCars(page: number, limit = 7) {
+  async getCars(page: number, limit = 7): Promise<{ items: ICar[], count: string | null }> {
     const response = await fetch(`${this.garage}?_page=${page}&_limit=${limit}`);
-  const data: ICar[] = await response.json();
-  const count: string | null = await response.headers.get('X-Total-Count');
-  console.log({
-    items: data,
-    count: count
-  });
-  return {
+    const data: ICar[] = await response.json();
+    const countCars: string | null = await response.headers.get('X-Total-Count');
+    return {
       items: data,
-      count: count
+      count: countCars,
     };
   }
 }
